valnum, numformatter, wxWigetsWindowPlacement moved to lib-shuttlegui
