+++ libraries/lib-shuttlegui/numformatter.cpp
