+++ libraries/lib-shuttlegui/numformatter.h
