+++ libraries/lib-shuttlegui/valnum.cpp
