+++ libraries/lib-shuttlegui/wxWidgetsWindowPlacement.cpp
