+++ src/commands/CompareAudioCommand.cpp
@@ -34,7 +34,7 @@ threshold of difference in two selected tracks
 #include "SettingsVisitor.h"
 #include "ShuttleGui.h"
 #include "../widgets/AudacityMessageBox.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "CommandContext.h"
 
 const ComponentInterfaceSymbol CompareAudioCommand::Symbol
