+++ src/effects/Amplify.cpp
@@ -36,7 +36,7 @@
 
 #include "ShuttleGui.h"
 #include "WaveTrack.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 
 enum
