+++ src/effects/ChangePitch.cpp
@@ -38,7 +38,7 @@ the pitch without changing the tempo.
 #include "ShuttleGui.h"
 #include "Spectrum.h"
 #include "WaveTrack.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "TimeWarper.h"
 
 // Soundtouch defines these as well, which are also in generated configmac.h
