+++ src/effects/ChangeSpeed.cpp
@@ -29,7 +29,7 @@
 #include "ShuttleGui.h"
 #include "../SyncLock.h"
 #include "../widgets/NumericTextCtrl.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 #include "TimeWarper.h"
 #include "WaveClip.h"
