+++ src/effects/ClickRemoval.cpp
@@ -36,7 +36,7 @@
 #include "Prefs.h"
 #include "ShuttleGui.h"
 #include "../widgets/AudacityMessageBox.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 #include "WaveTrack.h"
