+++ src/effects/Distortion.cpp
@@ -41,7 +41,7 @@
 
 #include "Prefs.h"
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 const EnumValueSymbol EffectDistortion::kTableTypeStrings[nTableTypes] =
 {
