+++ src/effects/DtmfGen.cpp
@@ -25,7 +25,7 @@
 #include "Prefs.h"
 #include "ShuttleGui.h"
 #include "../widgets/NumericTextCtrl.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 
 namespace {
