+++ src/effects/Loudness.cpp
@@ -27,7 +27,7 @@
 #include "../ProjectFileManager.h"
 #include "ShuttleGui.h"
 #include "WaveTrack.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "../widgets/ProgressDialog.h"
 
 #include "LoadEffects.h"
