+++ src/effects/Noise.cpp
@@ -25,7 +25,7 @@
 
 #include "Prefs.h"
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "../widgets/NumericTextCtrl.h"
 
 const EnumValueSymbol EffectNoise::kTypeStrings[nTypes] =
