+++ src/effects/Normalize.cpp
@@ -29,7 +29,7 @@
 #include "../ProjectFileManager.h"
 #include "ShuttleGui.h"
 #include "WaveTrack.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "../widgets/ProgressDialog.h"
 
 const EffectParameterMethods& EffectNormalize::Parameters() const
