+++ src/effects/Phaser.cpp
@@ -28,7 +28,7 @@
 #include <wx/slider.h>
 
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 const EffectParameterMethods& EffectPhaser::Parameters() const
 {
