+++ src/effects/Repeat.cpp
@@ -34,7 +34,7 @@
 #include "WaveTrack.h"
 #include "WaveClip.h"
 #include "../widgets/NumericTextCtrl.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 #include "LoadEffects.h"
