+++ src/effects/TimeScale.cpp
@@ -25,7 +25,7 @@
 
 #include "MemoryX.h"
 #include "ShuttleGui.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 
 enum
 {
