+++ src/effects/TruncSilence.cpp
@@ -34,7 +34,7 @@
 #include "ShuttleGui.h"
 #include "../SyncLock.h"
 #include "WaveTrack.h"
-#include "../widgets/valnum.h"
+#include "valnum.h"
 #include "../widgets/AudacityMessageBox.h"
 
 class Enums {
