+++ src/effects/VST/VSTEffect.cpp
@@ -83,7 +83,7 @@
 #include "PlatformCompatibility.h"
 #include "../../SelectFile.h"
 #include "ShuttleGui.h"
-#include "../../widgets/valnum.h"
+#include "valnum.h"
 #include "../../widgets/AudacityMessageBox.h"
 #include "../../widgets/NumericTextCtrl.h"
 #include "XMLFileReader.h"
