+++ src/effects/lv2/LV2Validator.cpp
@@ -40,7 +40,7 @@
 
 #include <wx/scrolwin.h>
 
-#include "../../widgets/valnum.h"
+#include "valnum.h"
 
 #include "../../widgets/NumericTextCtrl.h"
