+++ src/effects/vamp/VampEffect.cpp
@@ -33,7 +33,7 @@
 
 
 #include "ShuttleGui.h"
-#include "../../widgets/valnum.h"
+#include "valnum.h"
 #include "../../widgets/AudacityMessageBox.h"
 
 #include "../../LabelTrack.h"
