+++ src/tracks/labeltrack/ui/LabelTrackView.cpp
@@ -38,7 +38,7 @@ Paul Licameli split from TrackPanel.cpp
 #include "UndoManager.h"
 #include "ViewInfo.h"
 #include "AudacityTextEntryDialog.h"
-#include "../../../widgets/wxWidgetsWindowPlacement.h"
+#include "wxWidgetsWindowPlacement.h"
 
 #include <wx/clipbrd.h>
 #include <wx/dcclient.h>
